Fix Maps displaying twice on game start
The GameUpdateEvent is being fired a little too much at the moment, probably broke stats as well. But I don't have time to fix that now.
The problem is that you can be sent from the same game, to the same game, to the same server. And there isn't really a way for me to check it. Maybe I should use the join message? But then stuff breaks for nicks, ugh... Will have to think about it